+++ HISTORY.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## 3.1.0 (23/04/2019)
+* When using `Wavefront::Write`, large numbers of points are written
+  in chunks, rather than all at once. The chunk size can be set by
+  the user when instantiating the class.
+
 ## 3.0.2 (06/04/2019)
 * Better handling of non-existent or malformed config files.
 * Look for `~/.wavefront.conf` as well as `~/.wavefront`. Both these
-  fixes are related to findind out that other Wavefront tooling
+  fixes are related to finding out that other Wavefront tooling
   creates `~/.wavefront` as a directory.
 
 ## 3.0.1 (05/04/2019)